camera_manager/requirements.txt
netmon
django-model-utils